# Performance improvement on Delta.Push